Exchange Rate Provider of the Federal Reserve Bank of the United States.
This uses the feed at
http://www.federalreserve.gov/feeds/data/H10_H10.XML to obtain the prior
weekdays' noon-time exchange rates.

Should this feature be included in a future version of the Moneta
JSR-354 RI implementation, I believe it would find broad utilization and
would be a welcome supplement to the existing ECB and IMF providers.  My
personal motivation in developing this was to obtain a reliable source
for TWD currency exchange rates, which are not available with the other
two providers.  Having this important data source readily available to
Moneta users will enhance the out-of-the-box user experience.

A unit test class is included.