Add retry mechanism to `CompoundRateProvider.getExchangeRate` on failure
Modified the getExchangeRate method in the CompoundRateProvider class to introduce a failFast parameter.

Previously, the method would immediately throw an exception on the first provider failure. This change allows for a more flexible approach by adding an option to continue attempting to get an exchange rate from subsequent providers if failFast is set to false.

- Added a new overloaded getExchangeRate method with a failFast boolean parameter.
- If failFast is true, the method retains the original behavior of failing immediately on the first exception.
- If failFast is false, the method logs a warning and continues to the next provider in case of an exception.

This update addresses some of the limitations described in issue #385.